+++ src/Components/ExpressionArray.php
@@ -23,6 +23,7 @@ class ExpressionArray extends Component
      * @param array      $options parameters for parsing
      *
      * @return Expression[]
+     * @throws \PhpMyAdmin\SqlParser\Exceptions\ParserException
      */
     public static function parse(Parser $parser, TokensList $list, array $options = [])
     {
@@ -104,12 +105,20 @@ class ExpressionArray extends Component
 
         --$list->idx;
 
+        if (is_array($ret)) {
+            $expr = $ret[count($ret) - 1]->expr;
+            if (preg_match('/\s*--\s.*$/', $expr, $matches)) {
+                $found = $matches[0];
+                $ret[count($ret) - 1]->expr = substr($expr, 0, strlen($expr) - strlen($found));
+            }
+        }
+
         return $ret;
     }
 
     /**
-     * @param ExpressionArray[] $component the component to be built
-     * @param array             $options   parameters for building
+     * @param Expression[] $component the component to be built
+     * @param array        $options   parameters for building
      *
      * @return string
      */