+++ src/Lexer.php
@@ -346,6 +346,39 @@ class Lexer extends Core
 
         // Saving the tokens list.
         $this->list = $list;
+
+        $this->solveAmbiguityOnStarOperator();
+    }
+
+    /**
+     * Resolves the ambiguity when dealing with the "*" operator.
+     *
+     * In SQL statements, the "*" operator can be an arithmetic operator (like in 2*3) or an SQL wildcard (like in
+     * SELECT a.* FROM ...). To solve this ambiguity, the solution is to find the next token, excluding whitespaces and
+     * comments, right after the "*" position. The "*" is for sure an SQL wildcard if the next token found is any of:
+     * - "FROM" (the FROM keyword like in "SELECT * FROM...");
+     * - "USING" (the USING keyword like in "DELETE table_name.* USING...");
+     * - "," (a comma separator like in "SELECT *, field FROM...");
+     * - ")" (a closing parenthesis like in "COUNT(*)").
+     * This methods will change the flag of the "*" tokens when any of those condition above is true. Otherwise, the
+     * default flag (arithmetic) will be kept.
+     *
+     * @return void
+     */
+    private function solveAmbiguityOnStarOperator()
+    {
+        $iBak = $this->list->idx;
+        while (null !== ($starToken = $this->list->getNextOfTypeAndValue(Token::TYPE_OPERATOR, '*'))) {
+            // ::getNext already gets rid of whitespaces and comments.
+            if (($next = $this->list->getNext()) !== null) {
+                if (($next->type === Token::TYPE_KEYWORD && in_array($next->value, ['FROM', 'USING'], true))
+                    || ($next->type === Token::TYPE_OPERATOR && in_array($next->value, [',', ')'], true))
+                ) {
+                    $starToken->flags = Token::FLAG_OPERATOR_SQL;
+                }
+            }
+        }
+        $this->list->idx = $iBak;
     }
 
     /**
@@ -842,6 +875,7 @@ class Lexer extends Core
      * @param string $quote additional starting symbol
      *
      * @return null|Token
+     * @throws LexerException
      */
     public function parseString($quote = '')
     {
@@ -889,6 +923,7 @@ class Lexer extends Core
      * Parses a symbol.
      *
      * @return null|Token
+     * @throws LexerException
      */
     public function parseSymbol()
     {
@@ -914,7 +949,7 @@ class Lexer extends Core
         $str = null;
 
         if ($this->last < $this->len) {
-            if (($str = $this->parseString('`')) === null) {
+            if (($str = $this->parseString(Context::getIdentifierQuote())) === null) {
                 if (($str = $this->parseUnknown()) === null) {
                     $this->error(
                         'Variable name was expected.',