+++ src/Parser.php
@@ -375,6 +375,7 @@ class Parser extends Core
 
     /**
      * Builds the parse trees.
+     * @throws ParserException
      */
     public function parse()
     {