+++ tests/Parser/ParserTest.php
@@ -28,6 +28,7 @@ class ParserTest extends TestCase
             ['parser/parse'],
             ['parser/parse2'],
             ['parser/parseDelimiter'],
+            ['parser/ansi/parseAnsi'],
         ];
     }