+++ tests/Utils/CLITest.php
@@ -96,14 +96,14 @@ class CLITest extends TestCase
             ],
             [
                 ['h' => true],
-                'Usage: highlight-query --query SQL [--format html|cli|text]' . "\n" .
+                'Usage: highlight-query --query SQL [--format html|cli|text] [--ansi]' . "\n" .
                 '       cat file.sql | highlight-query' . "\n",
                 0,
             ],
             [
                 [],
                 'ERROR: Missing parameters!' . "\n" .
-                'Usage: highlight-query --query SQL [--format html|cli|text]' . "\n" .
+                'Usage: highlight-query --query SQL [--format html|cli|text] [--ansi]' . "\n" .
                 '       cat file.sql | highlight-query' . "\n",
                 1,
             ],
@@ -161,7 +161,7 @@ class CLITest extends TestCase
             [
                 '',
                 ['h' => true],
-                'Usage: highlight-query --query SQL [--format html|cli|text]' . "\n" .
+                'Usage: highlight-query --query SQL [--format html|cli|text] [--ansi]' . "\n" .
                 '       cat file.sql | highlight-query' . "\n",
                 0,
             ],
@@ -169,7 +169,7 @@ class CLITest extends TestCase
                 '',
                 [],
                 'ERROR: Missing parameters!' . "\n" .
-                'Usage: highlight-query --query SQL [--format html|cli|text]' . "\n" .
+                'Usage: highlight-query --query SQL [--format html|cli|text] [--ansi]' . "\n" .
                 '       cat file.sql | highlight-query' . "\n",
                 1,
             ],
@@ -226,14 +226,14 @@ class CLITest extends TestCase
                 '',
                 [],
                 'ERROR: Missing parameters!' . "\n" .
-                'Usage: lint-query --query SQL' . "\n" .
+                'Usage: lint-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | lint-query' . "\n",
                 1,
             ],
             [
                 '',
                 ['h' => true],
-                'Usage: lint-query --query SQL' . "\n" .
+                'Usage: lint-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | lint-query' . "\n",
                 0,
             ],
@@ -292,14 +292,14 @@ class CLITest extends TestCase
             ],
             [
                 ['h' => true],
-                'Usage: lint-query --query SQL' . "\n" .
+                'Usage: lint-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | lint-query' . "\n",
                 0,
             ],
             [
                 [],
                 'ERROR: Missing parameters!' . "\n" .
-                'Usage: lint-query --query SQL' . "\n" .
+                'Usage: lint-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | lint-query' . "\n",
                 1,
             ],
@@ -345,14 +345,14 @@ class CLITest extends TestCase
             ],
             [
                 ['h' => true],
-                'Usage: tokenize-query --query SQL' . "\n" .
+                'Usage: tokenize-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | tokenize-query' . "\n",
                 0,
             ],
             [
                 [],
                 'ERROR: Missing parameters!' . "\n" .
-                'Usage: tokenize-query --query SQL' . "\n" .
+                'Usage: tokenize-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | tokenize-query' . "\n",
                 1,
             ],
@@ -396,7 +396,7 @@ class CLITest extends TestCase
             [
                 '',
                 ['h' => true],
-                'Usage: tokenize-query --query SQL' . "\n" .
+                'Usage: tokenize-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | tokenize-query' . "\n",
                 0,
             ],
@@ -404,7 +404,7 @@ class CLITest extends TestCase
                 '',
                 [],
                 'ERROR: Missing parameters!' . "\n" .
-                'Usage: tokenize-query --query SQL' . "\n" .
+                'Usage: tokenize-query --query SQL [--ansi]' . "\n" .
                 '       cat file.sql | tokenize-query' . "\n",
                 1,
             ],