+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 * Replace sscanf by equivalent native PHP functions because sscanf can be disabled for security reasons. (#270)
 * Allow phpunit 9
 * Fix for php error when "INSERT INTO x SET a = 1" is "INSERT INTO x SET = 1" (#295)
+* Fixed lexer fails to detect "*" as a wildcard (#288)
+* Fixed ANSI_QUOTES support (#284)
+* Fixed parser mistakes with comments (#156)
 
 ## [5.2.0] - 2020-01-07